Add support for installing "Managed Storage" manifests
This patch is the first step to support "Native Manifests", which has been
introduced in ESR60. To use this feature, you need to set up files as below:

    resources/
    ├── auto_confirm-0.6.xpi
    └── auto_confirm-0.6.xpi.ManagedStorage

The installer should automatically detect the existence of the manifest file
and distribute it to the appropriate place.

Right now, we only support "Managed Storage" manifests, but it should be
relatively trivial to add support for other manifests as well.